+++ CHANGELOG.md
@@ -6,6 +6,7 @@ GAMSPy 1.4.1
 - General
   - Fix implicit parameter validation bug.
   - Allow propagating bounds to the output variable in `flatten_dims` method.
+  - Migrate GAMSPy CLI to Typer.
 - Testing
   - Lower the number of dices in the interrupt test and put a time limit to the solve.
 - Documentation